packages/contracts-extractor/pyproject.toml
[project]
name = "context-router-contracts-extractor"
version = "0.7.0"
requires-python = ">=3.12"
description = "Cross-language service contract extraction (OpenAPI / protobuf / GraphQL) for context-router"
dependencies = [
    "context-router-contracts",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/contracts_extractor"]
